Logic fixes & cleanup of class
